backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2

# LLM providers (compatible OpenAI SDK)
openai>=1.40.0          # Groq utilise le même client OpenAI via base_url

# RAG
ragie>=1.0.0            # Ragie RAG (optionnel — fallback local si absent)
